wsd: net: move ServerSession to HttpServer.hpp
